🎨 폴더구조, 라우터 재설정
