+++ client/src/components/Common/HashBtn/HashBtn.tsx
