+++ client/src/components/Common/Header/SearchInput.tsx
