+++ client/src/components/Common/Header/UserArea.tsx
