+++ client/src/components/Features/Detail/Detail.tsx
