+++ client/src/components/Features/Mypage/type.ts
