+++ client/src/pages/Detail/Writecontent.tsx
