--- client/src/pages/Detailcontent.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Detailcontent = () => {
-  return <div>Detailcontent</div>;
-};
-
-export default Detailcontent;
